+++ docs/release-notes.md
@@ -7,6 +7,10 @@
   _If needed, you can update to SMAPI 3.16.0 first and then install the latest version._
 -->
 
+## Upcoming release
+* For players:
+  * Fixed logged SMAPI errors not having line numbers on Linux/macOS.
+
 ## 3.18.3
 Released 09 April 2023 for Stardew Valley 1.5.6 or later.